[MUNE-216] Add user controls
Updated the implementation of controls to better support matrix
and vector controls, and added the post_function feature to allow
algorithm-specific processing of parameter changes after the value
is set.

Added the peq algorithm, partly to test the post_function.

Added a configuration to use VB1 PEQ settings for playback.

Refactored the JACK code to clean it up a little bit.

Renamed things to get rid of the "test" nomenclature for things
that aren't really tests.  Unit tests should be done separately.