[MUNE-252] Improve file I/O usability
Previously, mune_dsp required JACK to be running at all times,
and had no graceful way of exiting for file-based processing.
Now it is possible to run file-based processing without JACK, and
to automatically quit (and close any output WAV files) after the
input WAV files are done playing.

If the configuration file contains any JACK input or output blocks,
the system will run indefinitely for real-time I/O.  Otherwise,
if any input WAV file blocks are present, the system will run
to the lengh of the longest input file and quit.  If no JACK
blocks or input WAV blocks are present, the system will run for
a time (in seconds) specified with the `--time` or `-t` command
line option.

Note that this requires the sample rate and frame size constants
to be specified for the session in the configuration file
(see `config/configuration.json` for example).

Additional command line options were added to specify the
configuration file (`--configuration`, `-c`) and parameter
defintion file (`--parameters`, `-p`), with the defaults being
`config/configuration.json` and `config/parameters.json`.

The default configuration is a simple file input and output with
a -6 dB gain applied.