[MUNE-255] Add comments about FFT scaling
Added comments to the class and `inverse()` FFT function to
emphasize that the N-point inverse FFT is not scaled by 1/N.
If a forward FFT is followed by an inverse FFT, the result will
need to be scaled by 1/N to match the original signal.

MATLAB and SciPy scale the inverse FFT by 1/N by default.  Other
options would be to scale only the forward FFT by 1/N, to scale
both the forward and inverse FFT by 1/sqrt(N), or to do no scaling
at all.

By leaving scaling out of the inverse FFT, we give the client code
the option of performing scaling anywhere.  Also, since windows
are usually used with forward/inverse FFTs, the windows could be
scaled to provide scaling without extra multiplications.