[MUNE-289] Refactor IIR filters
The IIR filter code was refactored to make it easier to reuse.
The coefficient and state memory is now allocated by an `IirFilter`
class.  This takes reduces code in the algorithms using IIR
filters, and hides some of the implementation-specific details,
such as the alignment of memory or the order of coefficients.

Various filter design methods are also integrated into the IIR
filter.  New design methods can easily be added and registered
for use by any algorithm that uses IIR filters.  See
`filter_design.cpp` for examples.

The filter design methods support filters that require multiple
second-order sections (such as a Butterworth LPF).  These have
not yet been tested.