[MUNE-248] Add profiling to DSP framework
Added profiling code to the DSP framework.  By default, each task
is profiled, but the individual blocks within the task are not.
Profiling each block within a task can be enabled by setting the
"profile_blocks" constant for that task to true.

The maximum time for each block may not be accurate because such
short time periods are being measured.  For best results, use a
configuration with many instances of the same type of block, and
use the results from the whole task.

There are a few restrictions and limitations with the current
implementation:

- The profiling results are only printed when the task is
    destroyed.  Running for a limited time (using the -t option)
    will cause the tasks to be destroyed after the given number
    of seconds.
- The results are only printed when using the JACK thread
    interface, rather than the callback interface.  Run mune_dsp
    with the -T option to enable the thread interface.