[MUNE-295] multi-threaded feedback suppression
The feedback suppression's analysis is performed once every 8192
samples, while the frame size is typically much smaller.  Running
this in the main thread causes the CPU load to be bursty.  The
analysis was moved to a lower-priority real-time thread to level
the processing load over time.  This reuses the `PeriodicTask`
class developed to support multi-threaded SIMULINK models.

- `include/bosepro/task.h`: add profiling to `PeriodicTask`
- `src/feedback_suppresion.cpp`: move analysis to a lower-priority
    real-time thread
- `src/iir.h`, `src/iir.cpp`: add a mutex to protect concurrent
    access of IIR coefficients when filter parameters are changed