[MUNE-295] Change FBS test frame size
Multi-threaded support was added for the FBS.  The current
implementation of `PeriodicTask` used for multi-threading
requires that the lower priority task is an integer multiple
of the higher priority task spawning it.  The original test
frame size for FBS was 480, which is commonly used in
conferencing applications due to VoIP packet sizes which are
usually 10 ms.  This is not compatible with the 8192 frame size
used for the FFT thread.  Changing the test configuration's frame
size to 512 works, and using power of 2 frame sizes is probably
easiest for pro audio applications where we aren't using VoIP.