[MUNE-93] Add tone control algorithm
Added the tone control EQ ("tone") algorithm.  This uses the same
design as ControlSpace, which has the low EQ using a 200 Hz low
shelf, the mid EQ using a 1 kHz peaking EQ, and the high EQ using
a 5 kHz high shelf.

Added the ControlSpace high shelf and low shelf designs to the
available IIR design methods.

Fixed an issue in the IIR filter that didn't count the number of
bands properly while processing the EQ.