[FUSION-45] Implement region allocation
Created a region memory allocation scheme using the C++17
`monotonic_buffer_resource`.  Different types of memory are
specified using a number DSP memory classes, such as
`DspStateMemory` or `DspSignalMemory`.  These operate in a way
similar to smart pointers, and specify the way in which the
algorithm expects to use the memory.  Each memory type can be
assigned to a memory region as appropriate to achieve the best
performance on a given platform.

Unit tests were added using doctest, and all run successfully.

This does not currently support creating and destroying new
regions: global regions are used for the entire execution of the
program.  When the capability to dynamically load and unload
circuits is added, we will need to enhance this code with the
ability to open and close new sets of regions for each thread.

The region memory features are not yet integrated into the
framework, and the algorithms have not been updated to use it.
This work will be done as part of FUSION-46, FUSION-47, and
FUSION-48.