FUSION-41 Remove smartmic algorithm
The "smartmic" algorithm was not really a functional algorithm,
a prototype to test integration between Simulink and the DSP
framework.  We do not plan on using Simulink moving forward, and
removing it will clean up the codebase.

All source files related to the algorithm were removed, and the
parameter definitions for the algorithm were removed.  The
Simulink generated code generated some warnings, which were
suppressed in `wscript`, and have now been re-enabled.