[FUSION-45] Add DSP memory to algorithm API
New versions of `assign_terminal()`, `assign_control()`, and
`assign_meter()` were created to support the use of `DspMemory`
to store values related to those parameters.  Corresponding
support within the `Terminal`, `Control`, and `Meter` classes was
also added.

The new versions help enforce that algorithms are using the
appropriate memory types for each paramter.

- Only `DspSignalMemory` can be assigned to terminals.
- Only `DspCoeffMemory` can be assigned to controls that don't use
    a post function.
- Only `DspParamMemory` can be assigned to controls that use a
    post function.
- Only `DspMeterMemory` can be assigned to meters.

The original versions of the assign functions (using
`std::vector`) have not yet been removed, but they will be
removed once all algorithms have been updated to use the new
memory allocation scheme.

The API for scalar controls and meters didn't change, because
they are genrally assigned pointers to member variables of
the algorithm class, which are not managed by DspMemory.

The `Algorithm` class's `meta` member (an instance of the
`AlgorithmMeta` class) was moved to `DspParamMemory`, since its
contents are never used in real time.