[FUSION-48] Use DspMemory for universal parameters
The processing of universal parameters for output signals (gain,
mute, bypass, and level meters) was updated to use DspMemory.
The data for output processing was moved to a separate class
(TerminalOutputProcessor) so that its data, which is used in real
time, can be stored in a fast memory region.  The Terminal data,
which is not used in real time, is stored separately.

All data internal to the TerminalOutputProcessor is stored in
DspStateMemory, DspCoeffMemory, or DspMeterMemory, as appropriate
for its associated use.  The API for assigning all parameters
associated with output processing has been updated to use the
DspMemory version of the API.

At the same time, all algorithms were updated to use
DspSignalMemory for all input and output signals.  It wasn't
practical to support both the new and old APIs (using std::vector
and std::unique_ptr for signals) at the same time.  The old APIs
for terminals were removed.  As part of this, the process()
function for IIR filters had to be updated to support
DspSignalMemory.

A defect in DspMemory related to destroying objects contained
in DspMemory when the DspMemory in destroyed was fixed.  It
previously assumed the objects would be created and destroyed
in the opposite order, which is not always the case.  The
destruction data is now stored in a map, which can be searced
for the correct destruction count.

Testing was performed by adding control changes for out_gain,
out_mute, and bypass in config/test_peq.json, and confirming that
the parameters behaved as expected.