[FUSION-47] Update algorithms for DspMemory
The old-style APIs for controls and meters (using std::vector)
were removed, and all algorithms were updated to use the
DspMemory version of the API.

Most algorithms (including the IIR and FFT) were updated to
internally use DspStateMemory, etc., as appropate for the usage
of each array.  In these algorithms, there is no remaining usage
of std::vector or std::unique_ptr for dynamically allocated
variables.

The existing config/test_*.json test configurations were used to
validate the algorithms work as expected.

In a few cases, algorithms have not been updated to use DspMemory
internally (though they still use it for terminals, controls, and
meters).  These algorithms are more complex, and will be updated
in separate stories.  These include:

- Arbitrary EQ
- Feedback Suppression
- ML Benchmark
- Percepnet

In addition, a defect was found in Arbitrary EQ, which is a
crash within `compute_coeffs()`, which will be addressed in a
separate story.