[FUSION-77] Fix ArbitraryEQ crash
The ArbitraryEQ algorithm was crashing during initialization as
the initial settings were being applied.  This turned out to be
the end of an array being overwritten by 1 element, which caused
an exception to be thrown when a `unique_ptr` went out of scope.
Reducing the length of the loop iterating over the array by 1
was enough to resolve the crash.  This was introduced before the
algorithm was ported to mune_dsp, but is the kind of bug that is
difficult to reproduce and detect.