[FUSION-85] Attach memory regions to tasks
Each `Task` class now has its own `RegionManager`, each with its
own set of memory regions.  Signal processing blocks and signals
within a task will use that task's region manager for all of their
memory allocation.  This allows each task to use a separate set
of regions, which will be destroyed when that task is destroyed.

Previously, there was a global region manager for all allocations,
which did not support dynamic loading and unloading of circuits.
Now, in the tasks's constructor, the task's region manager is
opened for blocks to allocate from.  The region manager is closed
at the end of the task's constructor, so that allocations from
outside of the task cannot be done from its memory manager.  The
region manager is opened again in the task's destructor, so that
object destruction counts can be retrieved from the regions while
arrays of objects managed by DspMemory can be destroyed.

Task creation and destruction will always be done from the same
non-real-time thread, so that a global pointer to the current
region manager to be used by DspMemory is sufficient.

Testing was done by running `config/test_peq.json` and ensuring
that no errors occurred.  A temporary log confirmed that all of
the task's regions were destroyed along with it (freeing up all
of its memory).  Future testing of multiple tasks in a design
will be done when the APIs for loading and unloading individual
tasks are created.