[FUSION-83] Add support for dynamic circuits.
Audio processing `Task` objects can now be created and destroyed
independently, while other tasks are running.  This provides
support for dynamically loadable circuits, though a "circuit" may
consist of multiple tasks, possibly on multiple devices, if the
circuit is too large to run on a single core.

Each task is associated with its own JACK client, and all "jack_in"
and "jack_out" will create JACK ports associated with that client.
The task can be connected to signals using these JACK ports to
connect to the JACK server system input and output ports, or to
ports in other tasks for circuits that operate on multiple tasks
on multiple cores.

Tasks can also be stopped and restarted without destroying them.
When stopped, they will remain initialized in memory, but their
JACK processing will be deactivated.  This will also disconnect
any JACK ports associated with the task.  When the task is
restarted, its ports will automatically be reconnected.

A very basic socket server was added to `mune_dsp` to allow
commands to be sent to the DSP for on-the-fly testing of
creating/destroying/starting/stopping tasks.  A simple PyQt UI
was developed (outside of this repo) for testing.  Testing
confirmed that tasks could be created and destroyed while the
system was running.

Summary of changes:

- Added start/stop functionality to tasks (`session.h`, `task.h`,
    `jack.h`, `jack.cpp`).
- Integrated task processing into JACK clients (`task.h`,
    `jack.h`, `jack.cpp`).
- Added JSON command processing (`property.h`, `configuration.h`,
    `session.h`, `task.h`).
- Added socket server (`mune_dsp.cpp`).
- Refactored management of JACK ports (`jack.h`, `jack.cpp`,
    `jack_in.cpp`, `jack_out.cpp`).