[FUSION-65] Add Variscite build
Support for building with the SDK and toolchain for the Variscite
i.MX 8M Mini development board was added.  Building for Variscite
can be configured using:

`source /opt/fsl-imx-xwayland/6.1-mickledore/environment`

which only needs to be done once per login, and:

`python3 waf configure --platform=varmini`

Doctest was moved to a submodule so that it could be included in
the build without having to install it as a prerequisite.  I
couldn't find a recipe to include doctest in the Yocto build for
Variscite, so a submodule seemed the best route.  This also
ensures that builds on all platforms are using the same version
of doctest, rather than what happens to be installed on the
system.