[FUSION-83] Fix crash on exit without JACK
The `Task` object was destroying its JACK client, even in
configurations where JACK is not used (i.e., testing with files).
This was crashing on exit.  Added a test to make sure the JACK
client pointer is valid before destroying the client, and a few
other similar tests where the `client` member was being
dereferenced.