[FUSION-116] Add meter transmission
Added the ability to transmit JSON-strings containing meter data
for all meters in the system.  To retrieve meters, a client sends
the command:

`{ "target": "session", "name": "send_meters" }`

This causes every block within every task to send JSON strings
(via a callback) for each of its meters.  The meter strings are
of the form:

`{ "source": "block_name", "name": "meter_name", "value": value }`

The value is correctly formatted JSON according to whether the
value is an integer, float, bool, or string.  For meters with
dimensions, a 1-D or 2-D array is created that contains all of
the meter values.  For example, the "out_meter" for a stereo
block might contain the value `[1.0, -3.5]`.

The meters were tested using the Python prototype UI, and the
build was successful for Mac, Linux, and the Variscite SDK.

We may want to allow the client to subscribe to particular meters
rather than receiving all of them.  This can be done later as a
lower priority.