[FUSION-248] Add task connections
Task connections are used to connect JACK ports between DSP tasks
(each running as a separate JACK client), and other DSP tasks,
other JACK clients, or JACK system I/O.

The ability to read task connections from the static configuration
was added.  Connection settings are applied to the JACK input and
output blocks, which make the connections once the JACK client
has started running (after initialization).

The JACK client name for a task now defaults to the task name, if
"jack_client_name" is not specified.  Similarly, the port name
prefix for a "jack_in" or "jack_out" block defaults to the name
of the block followed by an underscore, rather than defaulting to
"input_" or "output_" if the "port_name_prefix" property is not
specified.

To cause a task to be created without a JACK client, the
"is_jack_client" property (which defaults to true) of the task
can be set to false.  This should be done for configurations with
"wav_read" and "wav_write" file I/O and no "jack_in" or "jack_out"
blocks (such as for testing).  The test configurations that used
files were updated for this.

The "port_connection" parameter was removed from "jack_in" and
"jack_out", as the connections are now specified in the
"task_connections" list instead.

Testing was performed on the DRO output files for Prorotype 1,
and confirmed that the port connections were made in JACK.  The
file test configurations were also spot-checked.