[FUSION-153] Add signal generator
Updated the signal generator algorithm to include all of the
features specified.  It previously only supported a sine wave.

Added white and pink noise generator, uncluding a thread-safe
version of Donald Knuth's random number generator which is easier
to vectorize than the typical LCG.  The pink noise filter is based
on the signal generator in CSD.

Also added a sine sweep, which has a fast and slow sweep rate
using the same coefficients as CSD.