Incorporate fusion-services observer
Added the observer to receive parameter updates from the
fusion-services server.  Modified it slightly to create parameter
update JSON strings in the format used by the DSP.  Added a
command-line option (`--serverip`, `-s`) to specify the IP address
of the fustion server.

Added a test configuration for developers to test parameter
changes and hear the results.