FUSION-263 Use fusion-server given IP
If the Fusion server address is provided via the "serverip"
command-line option, the program will now connect to the
Fusion server, even if no JACK tasks are specified.  The DSP
will wait for the server to send commands (such as telling it
to load a configuration from a file).

Previously, the Fusion-server connection would only be made if
the program was started with a configuration that included
JACK tasks.

Also removed some defunct code for meters that became irrelevant
with the fusion server (soon to be repeaced by the new
telemetry/meter communication).