[FUSION-271] Fix JACK client early activation
The JACK client was being activated during its constructor.  This
caused the client to become active befor the audio task and all of
its blocks were finished initializing.  This was causing JACK to
fail with long initialization times (discovered when using the
serial terminal on a development board with verbose logs). Worse
yet, it can cause JACK to execute the thread's `process()` function
before the task is completely initialized.

It turned out the call to `jack_activate()` was redundant, as it
is also being called in `JackClient::start()`.  Removing it from
the `JackClient` constructor fixed the issue.  Some checks were
also added around port connections to avoid spurious error messages
when ports are connected before the client is activated.