[FUSION-271] Avoid error on task connections
A false error was being printed when JACK port connections are
specified between two tasks.  When the first task is started, it
tries to make all of its connections.  However, the second task's
JACK client is not yet active, preventing a successful connection.
An error message is printed, but then the connection is eventually
made when the second task is started.  Some checks were added to
avoid printing the error when the first task is started.