[FUSION-276] Fix task connection indexing bug
The task connections were incorrectly using the output channel
index to assign a task connection to a destination task, where
the input channel index for the destination block should be used.
This was a copy/paste error.  This input channel index is now
being used.

This was found and tested in the FUSION-267 branch.  It previously
caused a crash when it tried to assign a connection string to a bad
index, and now works as expected.