[FUSION-281] Load DSP config from server
The DSP configuration is now retrieved from the Fusion Server
whenever the static configuration is changed.  The front end sends
the configuration to the path "dsp_static_config".  The DSP will
receive an update that is the contents of the static configuration
from this path.  The DSP writes this to a file
(/tmp/dsp_config.json), and then loads the configuration from the
file (after first destroying any audio tasks that are currently
running).