[FUSION-282] Support updates of vector params
Updates of vector parameters in the Fusion Server are now
supported.  The DSP subscribes to the `"settings.audio.*.*[*]"`
path in addition to the `"settings.audio.*.*"` path.  Whenever
a vector parameter value is updated, the DSP receives this update
along with the index of the affected element.  If a whole vector
is updated, the DSP receives separate updates for each element
that changed.  Updates so matrix parameters from the Fusion Server
are not yet supported.