[FUSION-247] Add meter value conversions
Meter value conversions have been added so that meter values can
be translated from the internal representation of the algorithm
(such as linear gain) to a user-facing value (such as dB).
Currently only linear gain/amplitude to dB is used, but that is
by far the most common type of conversion.

A conversion function can now be specified when assigning
telemetry in the algorithm, and the conversion is applied to the
values as they are transmitted to the telemetry core.

In addition, meter minimum and maximum values as specified in
algorithm-definitions.json are now enforced when the meters are
transmitted.

The universal "out_meter" meters were updated with the value
conversion.  Also, some algorithms also either added the conversion
or implemented missing meters as appropriate.

This was tested on the Variscite board.