[FUSION-250] Update default configuration
The default configuration should be an empty configuration like
`base.json`, which waits for the fusion server to provide the
configuration to use at run time.  The default configuration
filename is `configuration.json`, which was just a simple gain
applied to a test WAV file, but not otherwise useful.

Moved `configuraiton.json` to `test_gain.json`, and then moved
`base.json` to `configuration.json`.