[FUSION-294] Temporarily bypass AGC by default
The AGC is introducing artifacts when music is processed.  We
need to investigate a fix for this (FUSION-295), but in the
meantime the AGC will be bypassed by default.  This will improve
the experience for the Prototype 1, which includes an AGC in the
signal path, but is not a central part of the demo.  When the AGC
performance is improved, it will no longer be bypassed by default.