[FUSION-292] Add version to algorithm defintions
A "version" field was added to the top level object of
`algorithm-definitions.json`.  Its value is currently "0.1.0" and
will always be a string conforming to Semantic Versioning
(<https://semver.org>).  Its intent is to help the front end and
other components that interact with the DSP track compatibility
as definitions change.