[FUSION-307] Fix server config crash
The DSP was crashing when it was loading a new configuration from
the server.  Now it correctly creates its audio tasks, loads any
parameter settings specified in the configuration, and makes any
JACK port connections specified.

This was tested on the variscite board with configurations being
sent from the front-end to the server.