[FUSION-267] Retry server initial state
If the fusion server is not up and running yet, the initial
call to `requestInitialState()` seems to be going in the bit
bucket, and the DSP never receives anything from the server.
Now, if the poll from the socket connected to the server
times out and the initial state has not been received, we
try the request again.  Currently, it seems to take the server
around 50 seconds after power on to respond to this request.
From then on, communication with the server proceeds normally.