[FUSION-312] Terminal default channels
There was a defect where the terminals weren't properly getting
the default value for a "channels" property, if such a property
was specified as the number of channels for a terminal, and the
property was not set in the configuration.  The code was trying
to retrieve the value from the terminal definition, rather than
the algorithm definition which contains the property definition.

Tested on Mac and the variscite board which had a design sent by
the front end that did not have the channels property value set
for a particular block.