[FUSION-293] Better error handling
There is now better error handling for unexpected static
configurations and dynamic parameter settings.  Errors cause an
exception to be thrown, and caught so that the DSP will refuse to
load a problematic configuration, or apply an improper parameter
setting, and then continue running until a new configuration is
loaded or a new parameter update.

Errors that are handled in this way include:

- Invalid JSON syntax
- Missing required fields
- Values that are of the wrong type, out of range, or otherwise
    not allowed
- Settings for parameters or blocks that don't exist
- Connections between terminals or blocks that don't exist

Due to the increased use of exceptions, some redundant error
messages and return values were removed.

Numerous unit tests were added to ensure exceptions were thrown
when invalid configurations or parameter settings were applied.