[FUSION-412] Rename from Mune to Fusion
While this code was started under the Mune project, that codename
is now defunct.  Since the code is now being developed for the
Fusion project, everything has been renamed.

This repo (fusion-dsp) was forked from the pro-mune-dsp repo, and
the original repo will be archived.  In this commit, all
references in code and documentation have been updated.  The build
scripts were also updated to reflect the new name for the binary
(fusion_dsp).