Add support for multiple ALSA formats
The ALSA I/O blocks will now attempt to open the ALSA device with
the highest possible quality, starting with floating-point, then
with PCM 32-, 24-, and finally 16-bit.  This allows the device
to be opened with the best quality available without having to
specify the sample format in the configuration.

Some other changes were added to facilitate working with the
FusionConnect LKM (and AES67 streams).