Add support for multiple devices
Previously, the DSP would look for a single-device static
configuration from the Fusion server on the path
"dsp_static_config".  Now it watches the path that will be
specified by the DRO for multiple device systems.
It watches "devices[*]" so that it gets a separate update for each
device that changes its static configuration.  If the "id" field
of the updated device doesn't match the device ID configured for
this instance of the DSP, it is ignored.  If the device ID
matches, the DSP loads the updated static configuration from the
"dsp_static_config" field of that device.

The device ID of the DSP instance can be specified at the command
line using the `-i` command-line option.  The device ID is a
string, currently set to a stringified index ("0", "1", etc.),
but may later be something like a UUID.

To facilitate running multiple instances on the same device
(to simulate a multi-device system on a single board), the DSP
can now be called with the `-n` command-line option to avoid
connecting to the telemetry core.

A small change was made to `wscript` to fix the macOS build.