Add retrieval of device ID through curl
The Observer now uses libcurl to grab the device ID of this device
from the Fusion Server from the path `localhost:9090/device`.
When the Observer receives an update for the "devices" path, which
contains the DSP static configurations for each device, it will
check the device ID to see if it has changed, and then implement
the DSP static configuration for this device.

Note that this requires `curl` to be included the BSP.  That
provides libcurl as well as the command-line utility.