Add source selector algorithm
Added the source selector output, which copies the selected input
to the output.

A new version of the `get_maximum_value()` function was added to
`Definition` to allow the maximum value to either be specified
explicitly as a value, or as the name of a property or terminal
from which the maximum value is derived.  In the case of the
source selector, this allows the maximum value of the "input"
parameter (which specifies the index of the input channel to use
as the source) to have its maximum value determined by the number
if channels in the "input" terminal.  The "delay" algorithm was
also modified so that the maximum value of its "delay" parameter
is determined by the value of the "max_delay" property.

Testing was performed with `config/test_source_selector.json` and
also with `config/test_delay.json`.