Add multiple JACK connections per port
JACK allows multiple connections to the same input or output port.
In the case of output ports, this allows you to connect one output
to multiple inputs.  In the case of input ports, JACK will mix all
of the connectings coming in.  The framework was only tracking one
connection per port for the purpose of connecting tasks to each
other.  Now a list of connections is kept for each port, allowing
multiple JACK connections to the same port to be specified in the
configuration and made at run time.