Search alsa devices by name
To allow the ALSA device to be opened using its stream ID, we have
to search through the list of devices in "hw:FusionConnect" to
find the device whose name matches the stream ID.  Once found, the
device number is then used to open the device (for example,
"hw:FusionConnect,1").

We had to do this because ALSA won't let you open a device name
like "hw:FusionConnect,FC_TX_49152".  Also, we tried creating
named devices in /etc/asound.conf, but those won't be recognized
by the DSP if the file is written after the DSP loads a design.