Fix AGC artifacts on music
The AGC was miscalculating when the current gain is close enough
to the target gain to stop moving the current gain.  This causes
modulation when the current gain is very close to the target gain,
and the target gain is not changing much.  This artifact is much
more obvious on music with steady tones than on speech. The
calculation of when the gains are close was corrected.

In addition, the AGC bypass parameter default was changed back to
`false`.  It was previously set to `true` due to this defect, but
is intended to be `false` in the algorithm specification.