ALSA close device with bad fd
When a Fusion or AES67 stream is torn down, the device disappears.
This causes all of the ALSA interactions with the device to
return -EBADFD.  When this occurs, we now close the device, and
go back to the state where we wait for a device to become
available.  This allows streams to come and go while the DSP is
still running.  It will also support devices like Bluetooth or
USB in the future.

In addition, the logs in the observer were converted to spdlog.