alsa-out-full
When the AlsaOut device is writing, if the output buffer depth
exceeds its target depth too much (i.e. the buffer is full), the
buffer was being fast forwarded rather than rewound.  In other
words, it was filling up more to compensate rather than emptying.
This would cause the device to be stuck in a permanently full
state. The sign of the buffer adjustment calculation was
corrected.