[FUSION-983] Fix limiter parameter units
The limiter peak and RMS attack and release times are specified in
milliseconds in the PO parameters and properties spec, but were
being interpreted in seconds in the code.  This was resulting in
unusably long attack and release times.  The parameters are now
scaled properly before being applied to coefficients.

In addition, the parameters were being written in their unconverted
form to the state variable, and shortly after were converted using
the update functions.  There is a small chance this could have
resulted in an incorrect value briefly, if the framework wrote the
intermediate value and was preempted by real-time processing before
the value was converted.  This would be very difficult to diagnose
in the field, and was fixed proactively.