[FUSION-997] Fix matrix crash
The storage for the coefficient memory for the "input_mute"
parameter in the matrix mixer algorith was defined as a
`DspCoeffMemory<bool *[]>` (a matrix) but should have been a
`DspCoeffMemory<bool []>` (a vector).  This was causing a crash
when a design including a matrix mixer was loaded.

The definition of the `input_mute` variable was corrected, and
error messages were added to the parameter `assign()` functions
in case the wrong dimension of memory is assigned to a paremter,
so that we have better bread crumbs to follow if this mistake
happens again in the future.

Tested on a C0 and verified that the error is printed if the
dimensions are wrong, and that the matrix doesn't cause a crash.