Implement paging router
The paging router has been implemented.  It functions similarly to
the ducker, but the paging inputs are mixed in with the non-paging
input.  There are a configurable number of paging inputs, in
increasing order of priority.  If a paging input is active, it
causes the default signal and all lower priority paging signals to
be ducked.

In addition, telemetry was modified to accept a named parameter to
define the maximum value of the telemetry (so that we can have the
"active_input" elemetry with the maximum value equal to the number
of paging inputs).