Add a minimum buffer depth for ALSA
With small frame sizes, the minimum and maximum buffer depths for
ALSA may be too tight to allow some drift in buffer depth due to
ASRC.

For the Proto2 demo, the minimum depth was 576 samples.  Before
this change, on C0 the minimum depth was 64 samples.  Increasing
this to 128 to start, but this can change to 256 or 512 as needed
(at the expense of more latency).