Add HDR ADC combiner algorithm
The HDR ADC combiner algorithm takes two channels of input: the
unity branch, which has unity analog gain (i.e., a full scale
analog input), and a gain branch which has some other analog gain
applied.  These are connected to "in" channels 1 and 2
respectively.  Depending on the signal levels, the combiner
selects between the two branches to provide a HDR output
(appearing on "out").

The "branch_gain" parameter specifies the amount of analog gain
applied before the ADC.  The "out_gain" parameter has an unusually
high range (+60 dB) so that sufficient gain for microphone inputs
can be applied.  If the output gain is set to 0, a full scale
input on the ADC in the unity branch is a floating-point +/- 1.0.

Tested on a Fusion-mini C0 for the HDR ADC demo.