Add DSP unshifter algorithm
The unshifter algorithm is a workaround for the case where a PCM
signal is right shifted by 1 bit.  Right shifting causes severe
distortion in the signal, as it wraps negative values to become
very large possible values.

The algorithm itself is simple: convert to a 32-bit integer, left
shift 1 bit, and convert back to float.

Tested on wav files that have right shifted by 1 bit, and
confirmed that the output is a normal signal.